requirements.txt
streamlit
pdfplumber
faiss-cpu
requests
openai
pandas
numpy
python-dotenv
